+++ profiler/src/main/java/com/navercorp/pinpoint/profiler/instrument/config/InstrumentConfig.java
@@ -34,7 +34,7 @@ public interface InstrumentConfig {
     String getApplicationServerType();
 
     int getCallStackMaxDepth();
-    short getCallStackMaxSequence();
+    int getCallStackMaxSequence();
     boolean isPropagateInterceptorException();
     String getProfileInstrumentEngine();
     boolean isSupportLambdaExpressions();
